feat: Add react-intersection-observer dependency for lazy loading images
This commit adds the "react-intersection-observer" dependency to the project. This library allows for lazy loading of images by detecting when they enter the viewport. By using this library, we can improve the page load performance by only loading images when they are actually visible to the user.

This change ensures that the project is using the latest stable version of "react-intersection-observer" (v9.13.1) and enables the implementation of lazy loading for images.