feat: Improve lazy loading for images in career and services pages
This commit updates the `page.tsx` file in the `app/career` and `app/services` directories to improve lazy loading for images.

In the `career/page.tsx` file, the `className` of the hero section has been modified to include `mt-5`, adjusting the top margin for better spacing. Additionally, the `className` of the h1 and p elements has been updated to improve the typography and spacing.

In the `services/businessi/page.tsx` and `services/datascience/page.tsx` files, the `className` of the feature cards and tech stack buttons has been modified to improve the layout and styling.

These changes enhance the user experience by optimizing the loading of images and improving the visual appearance of the career and services pages.