feat: Improve lazy loading for images in career and services pages
This commit optimizes lazy loading for images in the career and services pages. It updates the `globals.css` file to comment out the font-family declarations for 'Roboto' and 'Arial', which were not being used. Additionally, in the `career/page.tsx` file, a line break element `<br></br>` has been added to improve the spacing between paragraphs.

These changes enhance the performance and visual appearance of the career and services pages by optimizing the loading of images and improving the typography.